feat: implemented FromIterator and created TryFromIterator trait, renamed main folders aswell.
