+++ src/structures/byte.rs
@@ -1,5 +1,7 @@
 use std::{num::{ParseIntError, TryFromIntError}, str::FromStr};
 use thiserror::Error;
+use crate::traits::try_from_iter::TryFromIterator;
+
 use super::bits::Bits;
 
 pub struct Byte {
@@ -26,24 +28,32 @@ pub enum ByteCreationError {
 
 impl Byte {
     pub fn from_bits(bits: &Bits) -> Result<Self, ByteCreationError> {
-        Self::from_iter(bits.value())
+        Self::try_from_iter(
+            bits
+                .value()
+                .to_owned()
+        )
     }
 
-    pub fn from_iter(iter: &[u8]) -> Result<Self, ByteCreationError> {
+    pub fn value(&self) -> &u8 {
+        &self.value
+    }
+}
+
+impl TryFromIterator<u8> for Byte {
+    type Error = ByteCreationError;
+
+    fn try_from_iter<I: IntoIterator<Item = u8>>(iter: I) -> Result<Self, Self::Error> {
         Ok(Self {
             value: u8::from_str_radix(
                 iter
-                    .iter()
-                    .map(|&byte| byte.to_string())
+                    .into_iter()
+                    .map(|byte| byte.to_string())
                     .collect::<String>()
                     .as_str(),
                 2
             )?
         })
-    }
-
-    pub fn value(&self) -> &u8 {
-        &self.value
     }
 }
 
@@ -63,6 +73,6 @@ impl FromStr for Byte {
             )
         }
 
-        Ok(Self::from_iter(&vec)?)
+        Ok(Self::try_from_iter(vec)?)
     }
 }
