+++ src/traits/try_from_iter.rs
@@ -1,0 +1,6 @@
+
+pub trait TryFromIterator<T>: Sized {
+    type Error;
+
+    fn try_from_iter<I: IntoIterator<Item = T>>(iter: I) -> Result<Self, Self::Error>;
+}
